Time units are now parameterized.
Was it worth the effort? Unlikely.
Do I know more about the haskell type system now? Undoubtably.
It's pretty ugly right now.
I'm working on reducing the ugly to a minimal subset of ugly,
to see if I can clean it up in a setting that's easier to reason
about. (Or, failing that, run to #haskell with a minimal example
and more intelligent questions). Ignore the Halp.hs file now,
it's tracking the generation of the minimal subset of ugly.