Added formatting of 'extra' in Dates
This includes adding 'isAbove', 'isBelow', and 'isForMe'
to format chunks. Generally, chunks with level > 0 are for your
left element, chunks with level < 0 are for your right element,
and chunks with level == 0 are for the current Format element.

(Note that a 'invert :: Chunk -> Chunk' is also necessary for
when you pass a negative chunk to your right element.)